--- packages/manager/src/features/NotificationCenter/Notifications.tsx
@@ -1,18 +1,0 @@
-import * as React from 'react';
-
-import { useFormattedNotifications } from './NotificationData/useFormattedNotifications';
-import { NotificationSection } from './NotificationSection';
-
-export const Notifications = () => {
-  const notifications = useFormattedNotifications();
-
-  return (
-    <NotificationSection
-      content={notifications}
-      emptyMessage="No notifications to display."
-      header="Notifications"
-    />
-  );
-};
-
-export default React.memo(Notifications);
